requirements.txt
streamlit
streamlit-extras
python-dotenv
pandas
ccxt
statsmodels
ratelimit
TA_Lib-0.4.24-cp310-cp310-win_amd64.whl